requirements.txt
pyyaml==5.4.1
redis==3.5.3
mysql-connector-python~=8.0.25
urllib3==1.26.4
concurrent-log==1.0.1
